[DDW-1147] Implement acceptance tests for Daedalus Balance wallets (#1816)
* [DDW-1147] Introduce E2E tests for Daedalus Balance wallets

* [DDW-1147] Separate steps definitions regarding to features belongs to and tests improvements

* [DDW-1147] Add E2E test scenarion for entering wrong password in transfer funds wizard

* [DDW-1147] Restructure restore wallet fetures by action instead wallet types

* [DDW-1147] Remove unnecessary test

* [DDW-1147] Improve hardware wallet restoration e2e test

* [DDW-1147] Restructuring test scenarios and code improvements

* [DDW-1147] Code cleanup

* [DDW-1147] Run prettier

* [DDW-1147] Remove unnecessary @watch decorator

* [DDW-1147] Fix broken E2E test for Balance Wallet resync

* [DDW-1147] Transfer Funds E2E test improvements

* [DDW-1147] CHANGELOG update

* [DDW-1147] Merges v2-integration
