+++ tests/addresses/e2e/features/receive-money.feature
@@ -3,8 +3,8 @@ Feature: Receive money
 
   Background:
     Given I have completed the basic setup
-    And I have a "Test Wallet" wallet with funds
-    And I have the following wallets:
+    And I have a "Test Wallet" rewards wallet with funds
+    And I have the following "Rewards" wallets:
       | name          |
       | Target Wallet |
