+++ tests/navigation/e2e/features/navigate-general-settings-menu.feature
@@ -3,7 +3,7 @@ Feature: General Settings
 
   Background:
     Given I have completed the basic setup
-    And I have the following wallets:
+    And I have the following "Rewards" wallets:
       | name        |
       | Test Wallet |
