+++ tests/navigation/e2e/features/navigate-wallet-tabs.feature
@@ -3,7 +3,7 @@ Feature: Navigate Wallet Tabs
 
   Background:
     Given I have completed the basic setup
-    And I have the following wallets:
+    And I have the following "Rewards" wallets:
       | name        |
       | Test wallet |
