+++ tests/navigation/e2e/features/toggle-sidebar-submenus.feature
@@ -3,7 +3,7 @@ Feature: Toggle Sidebar Submenus
 
   Background:
     Given I have completed the basic setup
-    And I have the following wallets:
+    And I have the following "Rewards" wallets:
       | name        |
       | Test Wallet |
