+++ tests/nodes/e2e/features/data-layer-migration.feature
@@ -10,7 +10,7 @@ Feature: Data Layer Migration
     Then I should not see the Data Layer Migration screen
 
   Scenario: I do have wallets
-    Given I have the following wallets:
+    Given I have the following "Rewards" wallets:
     | name   |
     | Wallet |
     And I have selected English language
