+++ tests/nodes/e2e/features/wallet-settings-recovery-phrase-verification.feature
@@ -4,7 +4,7 @@ Feature: Wallet Settings - Recovery Phrase Verification
 
   Background:
     Given I have completed the basic setup
-    And I have the following wallets:
+    And I have the following "Rewards" wallets:
       | name    |
       | Wallet  |
