+++ tests/transactions/e2e/features/send-money-to-receiver.feature
@@ -3,12 +3,12 @@ Feature: Send Money to Receiver
 
   Background:
     Given I have completed the basic setup
-    And I have the following wallets:
+    And I have the following "Rewards" wallets:
       | name   |
       | first  |
 
   Scenario: User Sends Money to Receiver
-    Given I have a "Test Wallet" wallet with funds
+    Given I have a "Test Wallet" rewards wallet with funds
     And I am on the "Test Wallet" wallet "send" screen
     And I can see the send form
     When I fill out the send form with a transaction to "first" wallet:
