+++ tests/transactions/e2e/features/transactions-display.feature
@@ -8,8 +8,8 @@ Feature: Display wallet transactions
 
   Background:
     Given I have completed the basic setup
-    And I have a "Test Wallet" wallet with funds
-    And I have the following wallets:
+    And I have a "Test Wallet" rewards wallet with funds
+    And I have the following "Rewards" wallets:
       | name          |
       | Target Wallet |
