+++ tests/transactions/e2e/features/wallet-utxos.feature
@@ -5,7 +5,7 @@ Feature: Wallet - UTXOs
     Given I have completed the basic setup
 
   Scenario: Chart visible. Page title and description
-    Given I have a "Test Wallet" wallet with funds
+    Given I have a "Test Wallet" rewards wallet with funds
     When I am on the "Test Wallet" wallet "utxo" screen
     Then the "title" element renders the following text:
       | message                     |
@@ -14,7 +14,7 @@ Feature: Wallet - UTXOs
     And the UTXOs chart is visible
 
   Scenario: The wallet does not contain any UTxOs
-    Given I have the following wallets:
+    Given I have the following "Rewards" wallets:
       | name      |
       | NewWallet |
     When I am on the "NewWallet" wallet "utxo" screen
