+++ tests/wallets/e2e/features/wallets-ordering.feature
@@ -1,21 +1,18 @@
 @e2e
-Feature: Wallet Odering
+Feature: Wallets ordering
 
   Background:
     Given I have completed the basic setup
+    Given I have a "Balance Wallet" balance wallet
 
-  Scenario: Wallets ordering
-    Given I have created the following wallets:
+  Scenario: "Daedalus Balance" wallet is shown on the bottom of the list below "Daedalus Rewards" wallet in order of creation
+    And I have created the following "Rewards" wallets:
       | name     |
       | Wallet 1 |
       | Wallet 2 |
       | Wallet 3 |
       | Wallet 4 |
       | Wallet 5 |
-      | Wallet 6 |
-      | Wallet 7 |
-      | Wallet 8 |
-      | Wallet 9 |
     Then I should see the wallets in the following order:
       | name     |
       | Wallet 1 |
@@ -23,7 +20,4 @@ Feature: Wallet Odering
       | Wallet 3 |
       | Wallet 4 |
       | Wallet 5 |
-      | Wallet 6 |
-      | Wallet 7 |
-      | Wallet 8 |
-      | Wallet 9 |
+      | Balance Wallet |
