+++ tests/wallets/e2e/steps/wallets.js
@@ -1,14 +1,8 @@
 // @flow
 import { Given, When, Then } from 'cucumber';
 import { expect } from 'chai';
-import BigNumber from 'bignumber.js';
-import { DECIMAL_PLACES_IN_ADA } from '../../../../source/renderer/app/config/numbersConfig';
 import {
-  addWalletPage,
-  importWalletHelpers,
-  isActiveWalletBeingRestored,
   createWallets,
-  fillOutWalletSendForm,
   getWalletByName,
   waitUntilWalletIsLoaded,
   addOrSetWalletsForScenario,
@@ -16,30 +10,38 @@ import {
   restoreLegacyWallet,
   waitUntilUrlEquals,
   navigateTo,
-  i18n,
-  waitForActiveRestoreNotification,
 } from './helpers';
-import { waitUntilTextInSelector, scrollIntoView } from '../../../common/e2e/steps/helpers';
-import {
-  sidebarHelpers,
-} from '../../../navigation/e2e/steps/helpers';
 import type { Daedalus } from '../../../types';
 
 declare var daedalus: Daedalus;
 
-Given(/^I have a "([^"]*)" wallet with funds$/, async function(walletName) {
+// Create "Rewards" wallets
+Given(/^I have the following "Rewards" wallets:$/, async function(table) {
+  await createWallets(table.hashes(), this);
+});
+
+// Creates them sequentially
+Given(/^I have created the following "Rewards" wallets:$/, async function(table) {
+  await createWallets(table.hashes(), this, { sequentially: true });
+});
+
+Given(/^I have a "([^"]*)" rewards wallet with funds$/, async function(walletName) {
   await restoreWalletWithFunds(this.client, { walletName });
   const wallet = await waitUntilWalletIsLoaded.call(this, walletName);
   addOrSetWalletsForScenario.call(this, wallet);
 });
 
-Given(/^I have the following wallets:$/, async function(table) {
-  await createWallets(table.hashes(), this);
+// Create "Balance" wallets
+Given(/^I have a "([^"]*)" balance wallet$/, async function(walletName) {
+  await restoreLegacyWallet(this.client, { walletName, hasFunds: false });
+  const wallet = await waitUntilWalletIsLoaded.call(this, walletName);
+  addOrSetWalletsForScenario.call(this, wallet);
 });
 
-// Creates them sequentially
-Given(/^I have created the following wallets:$/, async function(table) {
-  await createWallets(table.hashes(), this, { sequentially: true });
+Given(/^I have a "([^"]*)" balance wallet with funds$/, async function(walletName) {
+  await restoreLegacyWallet(this.client, { walletName, hasFunds: true });
+  const wallet = await waitUntilWalletIsLoaded.call(this, walletName);
+  addOrSetWalletsForScenario.call(this, wallet);
 });
 
 Given(/^I am on the "([^"]*)" wallet "([^"]*)" screen$/, async function(
@@ -48,67 +50,6 @@ Given(/^I am on the "([^"]*)" wallet "([^"]*)" screen$/, async function(
 ) {
   const wallet = getWalletByName.call(this, walletName);
   await navigateTo.call(this, `/wallets/${wallet.id}/${screen}`);
-});
-
-Given(/^I see the add wallet page/, function() {
-  return addWalletPage.waitForVisible(this.client);
-});
-
-Given(/^I see delete wallet dialog$/, function() {
-  return this.client.waitForVisible('.DeleteWalletConfirmationDialog_dialog');
-});
-
-Given(/^I see the create wallet dialog$/, function() {
-  return this.client.waitForVisible('.WalletCreateDialog');
-});
-
-Given(/^I see the restore wallet dialog$/, function() {
-  return this.client.waitForVisible('.WalletRestoreDialog_component');
-});
-
-Given(/^I dont see the create wallet dialog(?: anymore)?$/, function() {
-  return this.client.waitForVisible('.WalletCreateDialog', null, true);
-});
-
-When(/^I click on the create wallet button on the add wallet page/, function() {
-  return this.waitAndClick('.WalletAdd .createWalletButton');
-});
-
-When(/^I click on the import wallet button on the add wallet page/, function() {
-  return addWalletPage.clickImportButton(this.client);
-});
-
-When(/^I see the import wallet dialog$/, function() {
-  return importWalletHelpers.waitForDialog(this.client);
-});
-
-When(/^I select a valid wallet import key file$/, function() {
-  this.waitAndClick('.WalletFileImportDialog .FileUploadWidget_dropZone');
-});
-
-When(/^I enter wallet spending password:$/, async function(table) {
-  const fields = table.hashes()[0];
-  await this.client.setValue(
-    '.WalletFileImportDialog .spendingPassword input',
-    fields.password
-  );
-  await this.client.setValue(
-    '.WalletFileImportDialog .repeatedPassword input',
-    fields.repeatedPassword
-  );
-});
-
-When(
-  /^I click on the import wallet button in import wallet dialog$/,
-  function() {
-    return importWalletHelpers.clickImport(this.client);
-  }
-);
-
-When(/^I should see wallet spending password inputs$/, function() {
-  return this.client.waitForVisible(
-    '.WalletFileImportDialog .spendingPassword input'
-  );
 });
 
 When(/^I have one wallet address$/, function() {
@@ -122,81 +63,10 @@ When(/^I enter spending password "([^"]*)"$/, function(password) {
   );
 });
 
-When(
-  /^I click on the restore wallet button on the add wallet page$/,
-  function() {
-    return this.waitAndClick('.WalletAdd .restoreWalletButton');
-  }
-);
-
 When(/^I click the wallet (.*) button$/, async function(buttonName) {
   const buttonSelector = `.NavButton_component.${buttonName}`;
   await this.client.waitForVisible(buttonSelector);
   await this.client.click(buttonSelector);
-});
-
-When(/^I can see the send form$/, function() {
-  return this.client.waitForVisible('.WalletSendForm');
-});
-
-When(/^I fill out the wallet send form with:$/, function(table) {
-  return fillOutWalletSendForm.call(this, table.hashes()[0]);
-});
-
-When(
-  /^I fill out the send form with a transaction to "([^"]*)" wallet:$/,
-  async function(walletName, table) {
-    const values = table.hashes()[0];
-    const walletId = getWalletByName.call(this, walletName).id;
-    const walletAddress = await this.client.executeAsync((id, done) => {
-      daedalus.api.ada
-        .getAddresses({ walletId: id, isLegacy: false })
-        .then(response => done(response[0].id))
-        .catch(error => done(error));
-    }, walletId);
-    values.address = walletAddress.value;
-    return fillOutWalletSendForm.call(this, values);
-  }
-);
-
-When(/^the transaction fees are calculated$/, async function() {
-  this.fees = await this.client.waitUntil(async () => {
-    // Expected transactionFeeText format "+ 0.000001 of fees"
-    const transactionFeeText = await this.client.getText(
-      '.AmountInputSkin_fees'
-    );
-    const transactionFeeAmount = new BigNumber(transactionFeeText.substr(2, 8));
-    return transactionFeeAmount.greaterThan(0) ? transactionFeeAmount : false;
-  });
-});
-
-When(/^I click on the next button in the wallet send form$/, async function() {
-  const submitButton = '.WalletSendForm_nextButton';
-  await this.client.waitForVisible(submitButton);
-  return this.client.click(submitButton);
-});
-
-When(/^I see send money confirmation dialog$/, function() {
-  return this.client.waitForVisible('.WalletSendConfirmationDialog_dialog');
-});
-
-When(
-  /^I enter wallet spending password in confirmation dialog "([^"]*)"$/,
-  async function(password) {
-    await this.client.setValue(
-      '.WalletSendConfirmationDialog_passphrase input',
-      password
-    );
-  }
-);
-
-When(/^I submit the wallet send form$/, async function() {
-  await this.client.waitForEnabled(
-    '.WalletSendConfirmationDialog_dialog .confirmButton'
-  );
-  return this.client.click(
-    '.WalletSendConfirmationDialog_dialog .confirmButton'
-  );
 });
 
 When(
@@ -211,78 +81,6 @@ When(
   }
 );
 
-When(
-  /^I submit the create wallet with spending password dialog with the following inputs:$/,
-  async function(table) {
-    const fields = table.hashes()[0];
-    await this.client.setValue(
-      '.WalletCreateDialog .walletName input',
-      fields.walletName
-    );
-    await this.client.setValue(
-      '.WalletCreateDialog .spendingPassword input',
-      fields.password
-    );
-    await this.client.setValue(
-      '.WalletCreateDialog .repeatedPassword input',
-      fields.repeatedPassword
-    );
-    return this.client.click('.WalletCreateDialog .primary');
-  }
-);
-
-When(/^I enter wallet name "([^"]*)" in restore wallet dialog$/, async function(
-  walletName
-) {
-  return this.client.setValue(
-    '.ConfigurationDialog_input.walletName input',
-    walletName
-  );
-});
-
-When(/^I clear the recovery phrase in restore wallet dialog$/, async function() {
-  const words = await this.client.elements('.SimpleAutocomplete_selectedWordRemoveButton');
-  for (let i = words.value.length-1; i > -1; i--) {
-    const wordId = words.value[i].ELEMENT;
-    await this.client.elementIdClick(wordId);
-  }
-});
-
-When(/^I enter recovery phrase in restore wallet dialog:$/, async function(
-  table
-) {
-  const fields = table.hashes()[0];
-  const recoveryPhrase = fields.recoveryPhrase.split(' ');
-  for (let i = 0; i < recoveryPhrase.length; i++) {
-    const word = recoveryPhrase[i];
-    await this.client.setValue(
-      '.AutocompleteOverrides_autocompleteWrapper input',
-      word
-    );
-    await this.client.waitForVisible(`//li[text()="${word}"]`);
-    await this.waitAndClick(`//li[text()="${word}"]`);
-    await this.client.waitForVisible(`//span[text()="${word}"]`);
-  }
-});
-
-When(/^I enter wallet password in restore wallet dialog:$/, async function(
-  table
-) {
-  const fields = table.hashes()[0];
-  await this.client.setValue(
-    '.spendingPassword input',
-    fields.password
-  );
-  await this.client.setValue(
-    '.repeatPassword input',
-    fields.repeatedPassword
-  );
-});
-
-When(/^I submit the restore wallet dialog$/, function() {
-  return this.client.click('.WalletRestoreDialog .primary');
-});
-
 When(/^I click continue$/, function() {
   return this.waitAndClick('.primary');
 });
@@ -291,230 +89,6 @@ When(/^I click close$/, function() {
   return this.waitAndClick('.primary');
 });
 
-When(/^I click Check recovery phrase button$/, function() {
-  return this.waitAndClick('.primary');
-});
-
-When(/^I see the create wallet privacy dialog$/, function() {
-  return this.client.waitForVisible('.WalletBackupPrivacyWarningDialog');
-});
-
-When(
-  /^I click on "Please make sure nobody looks your screen" checkbox$/,
-  function() {
-    return this.waitAndClick(
-      '.WalletBackupPrivacyWarningDialog .SimpleCheckbox_root'
-    );
-  }
-);
-
-When(/^I submit the create wallet privacy dialog$/, function() {
-  return this.waitAndClick('.WalletBackupPrivacyWarningDialog .primary');
-});
-
-When(/^I see the create wallet recovery phrase display dialog$/, function() {
-  return this.client.waitForVisible('.WalletRecoveryPhraseDisplayDialog');
-});
-
-When(/^I note down the recovery phrase$/, async function() {
-  const recoveryPhrase = await this.client.getText(
-    '.WalletRecoveryPhraseMnemonic_component'
-  );
-  this.recoveryPhrase = recoveryPhrase.split(' ');
-});
-
-When(/^I submit the create wallet recovery phrase display dialog$/, function() {
-  return this.waitAndClick('.WalletRecoveryPhraseDisplayDialog .primary');
-});
-
-When(/^I see the create wallet recovery phrase entry dialog$/, function() {
-  return this.client.waitForVisible('.WalletRecoveryPhraseEntryDialog');
-});
-
-When(
-  /^I click on recovery phrase mnemonics in correct order$/,
-  async function() {
-    for (let i = 0; i < this.recoveryPhrase.length; i++) {
-      const word = this.recoveryPhrase[i];
-      const selector = 'MnemonicWord_root';
-      const disabledSelector = 'MnemonicWord_disabled';
-      await this.waitAndClick(
-        `//button[contains(@class,'${selector}') and not(contains(@class, '${disabledSelector}')) and text()="${word}"]`
-      );
-    }
-  }
-);
-
-When(/^I click on the "Accept terms" checkboxes$/, async function() {
-  const termsCheckboxes = await this.client.elements('.SimpleCheckbox_root');
-  for (let i = 0; i < termsCheckboxes.value.length; i++) {
-    const termsCheckbox = termsCheckboxes.value[i].ELEMENT;
-    await this.client.elementIdClick(termsCheckbox);
-  }
-});
-
-When(/^I submit the create wallet recovery phrase entry dialog$/, function() {
-  return this.waitAndClick('.WalletRecoveryPhraseEntryDialog .primary');
-});
-
-When(/^I click on delete wallet button$/, async function() {
-  return this.waitAndClick('.WalletSettings_deleteWalletBox button');
-});
-
-When(/^I enter "([^"]*)" as name of the wallet to confirm$/, async function(
-  walletName
-) {
-  return this.client.setValue(
-    '.DeleteWalletConfirmationDialog_confirmationInput input',
-    walletName
-  );
-});
-
-When(
-  /^I click on the "Make sure you have access to backup before continuing" checkbox$/,
-  function() {
-    return this.waitAndClick(
-      '.DeleteWalletConfirmationDialog_dialog .SimpleCheckbox_root'
-    );
-  }
-);
-
-When(/^I submit the delete wallet dialog$/, function() {
-  return this.client.click('.DeleteWalletConfirmationDialog_dialog .primary');
-});
-
-When(/^I try to import the wallet with funds again$/, async function() {
-  await sidebarHelpers.activateCategory(this.client, { category: 'wallets' });
-  await sidebarHelpers.clickAddWalletButton(this.client);
-  await addWalletPage.waitForVisible(this.client);
-  await addWalletPage.clickImportButton(this.client);
-  this.waitAndClick('.WalletFileImportDialog .FileUploadWidget_dropZone');
-  this.waitAndClick('.Dialog_actions button');
-});
-
-Given(/^I restore "([^"]*)" balance wallet with funds$/, async function(walletName) {
-  await restoreLegacyWallet(this.client, { walletName });
-  const wallet = await waitUntilWalletIsLoaded.call(this, walletName);
-  addOrSetWalletsForScenario.call(this, wallet);
-});
-
-Then(
-  /^I should see section "([^"]*)"$/,
-  async function(text) {
-    await waitUntilTextInSelector(this.client, {
-      selector: '.WalletRestoreDialog_component .Dialog_content > div:last-child .RadioSet_label',
-      text,
-    });
-  }
-);
-
-Then(
-  /^I should see a screen titled "([^"]*)"$/,
-  async function(text) {
-    await waitUntilTextInSelector(this.client, {
-      selector: '.Dialog_title h1',
-      text,
-      ignoreCase: true
-    });
-  }
-);
-
-Then(
-  /^I click on option "([^"]*)"$/,
-  async function(text) {
-    await this.waitAndClick(`//*[contains(text(), "${text}")]/..`);
-  }
-);
-
-Then(
-  /^I confirm "([^"]*)"$/,
-  async function(text) {
-    const targetSelector = `//label[contains(text(), "${text}")]`;
-    await this.client.waitForVisible(targetSelector);
-    await scrollIntoView(this.client, targetSelector);
-    await this.client.click(targetSelector);
-  }
-);
-
-When(/^I click "12 words - Balance wallet" radio button/, function() {
-  return this.waitAndClick('.RadioSet_radiosContainer div:nth-child(1) label');
-});
-
-When(/^I click on "Rewards wallet" radio button/, function() {
-  return this.waitAndClick('.RadioSet_radiosContainer div:nth-child(2) label');
-});
-
-Then(
-  /^I see the import wallet dialog with an error that the wallet already exists$/,
-  async function() {
-    return importWalletHelpers.expectError(this.client, {
-      error: await i18n.formatMessage(this.client, {
-        id: 'api.errors.WalletAlreadyImportedError',
-      }),
-    });
-  }
-);
-
-Then(
-  /^I should not see the create wallet recovery phrase entry dialog anymore$/,
-  function() {
-    return this.client.waitForVisible(
-      '.WalletRecoveryPhraseEntryDialog',
-      null,
-      true
-    );
-  }
-);
-
-Then(/^I should not see the delete wallet dialog anymore$/, function() {
-  return this.client.waitForVisible(
-    '.DeleteWalletConfirmationDialog_dialog',
-    null,
-    true
-  );
-});
-
-Then(/^I should not see the import wallet dialog anymore$/, function() {
-  return importWalletHelpers.waitForDialog(this.client, { isHidden: true });
-});
-
-Then(/^I should not see the restore wallet dialog anymore$/, function() {
-  return this.client.waitForVisible('.WalletRestoreDialog', null, true);
-});
-
-Then(
-  /^I should see the restore status notification while import is running$/,
-  async function() {
-    // Only check the rendered DOM if the restore is still in progress
-    if (await isActiveWalletBeingRestored(this.client)) {
-      await waitForActiveRestoreNotification(this.client);
-    }
-  }
-);
-
-Then(
-  /^I should not see the restore status notification once import is finished$/,
-  async function() {
-    await waitForActiveRestoreNotification(this.client, { isHidden: true });
-  }
-);
-
-Then(
-  /^I should see the restore status notification while restore is running$/,
-  async function() {
-    // Only check the rendered DOM if the restore is still in progress
-    if (await isActiveWalletBeingRestored(this.client)) {
-      await waitForActiveRestoreNotification(this.client);
-    }
-  }
-);
-
-Then(
-  /^I should not see the restore status notification once restore is finished$/,
-  async function() {
-    await waitForActiveRestoreNotification(this.client, { isHidden: true });
-  }
-);
 
 Then(/^I should have newly created "([^"]*)" wallet loaded$/, async function(
   walletName
@@ -547,53 +121,8 @@ Then(/^I should be on the "([^"]*)" wallet "([^"]*)" screen$/, async function(
   return waitUntilUrlEquals.call(this, `/wallets/${wallet.id}/${screenName}`);
 });
 
-Then(/^"([^"]*)" wallet should have "([^"]*)" as id$/, async function(
-  walletName,
-  walletId
-) {
-  const wallet = getWalletByName.call(this, walletName);
-  expect(wallet.id).to.equal(walletId);
-});
-
 Then(/^I should be on the "([^"]*)" screen$/, async function(screenName) {
   return waitUntilUrlEquals.call(this, `/${screenName}`);
-});
-
-Then(
-  /^I should see the following error messages on the wallet send form:$/,
-  async function(data) {
-    const errorSelector = '.WalletSendForm_component .SimpleFormField_error';
-    await this.client.waitForText(errorSelector);
-    let errorsOnScreen = await this.client.getText(errorSelector);
-    if (typeof errorsOnScreen === 'string') errorsOnScreen = [errorsOnScreen];
-    const errors = data.hashes();
-    for (let i = 0; i < errors.length; i++) {
-      const expectedError = await this.intl(errors[i].message);
-      expect(errorsOnScreen[i]).to.equal(expectedError);
-    }
-  }
-);
-
-// TODO: refactor this to a less hackish solution (fees cannot easily be calculated atm)
-Then(/^the latest transaction should show:$/, async function(table) {
-  const expectedData = table.hashes()[0];
-  await this.client.waitForVisible('.Transaction_title');
-  let transactionTitles = await this.client.getText('.Transaction_title');
-  transactionTitles = [].concat(transactionTitles);
-  const expectedTransactionTitle = await this.intl(expectedData.title, {
-    currency: 'Ada',
-  });
-  expect(expectedTransactionTitle).to.equal(transactionTitles[0]);
-  let transactionAmounts = await this.client.getText('.Transaction_amount');
-  transactionAmounts = [].concat(transactionAmounts);
-  // Transaction amount includes transaction fees so we need to
-  // substract them in order to get a match with expectedData.amountWithoutFees.
-  // NOTE: we use "add()" as this is outgoing transaction and amount is a negative value!
-  const transactionAmount = new BigNumber(transactionAmounts[0]);
-  const transactionAmountWithoutFees = transactionAmount
-    .add(this.fees)
-    .toFormat(DECIMAL_PLACES_IN_ADA);
-  expect(expectedData.amountWithoutFees).to.equal(transactionAmountWithoutFees);
 });
 
 // Extended timeout is used for this step as it takes more than DEFAULT_TIMEOUT
@@ -616,34 +145,8 @@ Then(
 );
 
 Then(
-  /^I should see newly generated address as active address on the wallet receive screen$/,
+  /^"Balance" wallet badge should be visible in the wallet sidebar$/,
   async function() {
-    return this.client.waitUntil(async () => {
-      const activeAddress = await this.client.getText('.WalletReceive_hash');
-      const generatedAddress = await this.client.getText(
-        '.receiveAddress-1 .Address_addressId'
-      );
-      return generatedAddress === activeAddress;
-    });
+    return this.client.waitForVisible('.SidebarWalletMenuItem_active .LegacyBadge_component');
   }
 );
-
-Then(/^I should see the wallets in the following order:$/, async function(
-  table
-) {
-  const expectedWallets = table.hashes();
-  const wallets = await this.client.getText('.SidebarWalletMenuItem_title');
-  wallets.forEach((wallet, index) =>
-    expect(wallet).to.equal(expectedWallets[index].name)
-  );
-});
-
-Given(/^I go back to the previous step$/, function() {
-  return this.waitAndClick('.DialogBackButton_component');
-});
-
-Then(/^The error message should be (hidden|visible)$/, function(state) {
-  const isVisible = state === 'visible';
-  return this.client.waitForVisible('.ConfigurationDialog_error', null, !isVisible);
-});
-
